refactor(cli): decouple reconcile from server
Move the reconcile REST wire contract into harness-protocol and use it from both the server route and CLI client. This lets reconcile compile without the server feature while keeping serve as the only in-process server command.
