+++ crates/harness-server/src/http/rest_contract.rs
@@ -195,7 +195,6 @@ register_legacy_dtos!(
     serde_json::Value,
     crate::handlers::operator_monitor::OperatorMonitorResponse,
     crate::handlers::projects::RegisterProjectRequest,
-    crate::handlers::reconcile::ReconcileParams,
     crate::handlers::runtime_hosts::RegisterRuntimeHostRequest,
     crate::handlers::runtime_project_cache::SyncWatchedProjectsRequest,
     crate::handlers::usage_monitor::UsageMonitorQuery,
@@ -216,7 +215,6 @@ register_legacy_dtos!(
     crate::http::workflow_routes::IssueWorkflowByIssueQuery,
     crate::http::workflow_routes::IssueWorkflowByPrQuery,
     crate::http::workflow_routes::ProjectWorkflowByProjectQuery,
-    crate::reconciliation::ReconciliationReport,
     crate::workflow_runtime_submission::CreateTaskRequest,
     harness_core::agent::ApprovalDecision,
     harness_core::proof_of_work::ProofOfWork,
